move EigenDeviceConverter to device_context.h
