+++ paddle/platform/device_context_test.cc
@@ -24,7 +24,7 @@ TEST(Device, Init) {
   for (int i = 0; i < count; i++) {
     DeviceContext* device_context = new CUDADeviceContext(GPUPlace(i));
     Eigen::GpuDevice* gpu_device =
-        device_context->template get_eigen_device<Eigen::GpuDevice>();
+        device_context->template get_eigen_device<GPUPlace>();
     ASSERT_NE(nullptr, gpu_device);
     delete device_context;
   }
